% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bk_pm10_ave}
\alias{bk_pm10_ave}
\title{PM10 levels for sensors in and around Bangkok, Thailand}
\format{
A \code{sf} format POINT dataset with 68 observations and 1 field in WGS84 CRS (EPSG: 4326)
\describe{
\item{ave_pm10}{the average PM10 level at the sensor}
}
}
\source{
Created by Chris Brunsdon
}
\usage{
bk_pm10_ave
}
\description{
A point dataset of average readings of PM10 levels for 77 sensors in and around Bangkok downloaded on 24 April 2024 from Open AQ (\url{https://openaq.org}) who have an API (\url{https://api.openaq.org/})
}
\examples{
data(bk_pm10_ave)
}
\keyword{datasets}
