% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bk_prov}
\alias{bk_prov}
\title{PM10 levels for sensors in and around Bangkok, Thailand}
\format{
A \code{sf} format Polygon dataset with 1 area and 2 fields in WGS84 CRS (EPSG: 4326)
\describe{
\item{ADM1_EN}{the area name in English}
\item{ADM1_TH}{the area name in Thai}
}
}
\source{
Created by Chris Brunsdon
}
\usage{
bk_prov
}
\description{
A boundary of the Bangkok administrative area downloaded on 24 April 2024 from The Humanitarian Data Exchange (\url{https://data.humdata.org}) who have a site for Subnational Administrative Boundaries in Thailand (\url{https://data.humdata.org/dataset/cod-ab-tha})
}
\examples{
data(bk_prov)
}
\keyword{datasets}
