% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{columbus}
\alias{columbus}
\title{Columbus OH spatial analysis data set (from the spData package)}
\format{
A \code{sf} format Polygon dataset with 49 area and 22 fields in NAD83(NSRS2007) / UTM zone 17N projection (EPSG: 3724), attribute descriptions lifted from the spData descriptions, with all source data files prepared by Luc Anselin (Spatial Analysis Laboratory, Department of Agricultural and Consumer Economics, University of Illinois, Urbana-Champaign, url{http://sal.agecon.uiuc.edu/datasets/columbus.zip}).
\describe{
\item{AREA}{computed by ArcView}
\item{PERIMETER}{computed by ArcView}
\item{COLUMBUS_}{internal polygon ID (ignore)}
\item{COLUMBUS_I}{another internal polygon ID (ignore)}
\item{POLYID}{yet another polygon ID}
\item{NEIG}{neighborhood id value (1-49); conforms to id value used in Spatial Econometrics book}
\item{HOVAL}{housing value (in 1,000 USD)}
\item{INC}{household income (in 1,000 USD)}
\item{CRIME}{residential burglaries and vehicle thefts per thousand households in the neighborhood}
\item{OPEN}{open space in neighborhood}
\item{PLUMB}{percentage housing units without plumbing}
\item{DISCBD}{distance to CBD}
\item{X}{x coordinate (in arbitrary digitizing units, not polygon coordinates)}
\item{Y}{y coordinate (in arbitrary digitizing units, not polygon coordinates)}
\item{NSA}{north-south dummy (North=1)}
\item{NSB}{north-south dummy (North=1)}
\item{EW}{east-west dummy (East=1)}
\item{CP}{core-periphery dummy (Core=1)}
\item{THOUS}{constant=1,000}
\item{NEIGNO}{NEIG+1,000, alternative neighborhood id value}
}
}
\source{
Anselin, Luc. 1988. Spatial econometrics: methods and models. Dordrecht: Kluwer Academic, Table 12.1 p. 189.
}
\usage{
columbus
}
\description{
Neighbourhoods in Columbus, USA, with 1980 data attributes.
}
\examples{
data(columbus)
}
\keyword{datasets}
