% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crimes}
\alias{crimes}
\title{A point dataset of the locations of street crimes in the Leeds (UK) area in March 2021}
\format{
A \code{sf} format POINT object with  with 9180 observations (crimes) and 5 attributes in WGS84 projection (EPSG: 4326)
\describe{
\item{category}{the type of crime}
\item{street_name}{the location of the crime on the street}
\item{location_type}{either \code{Force} which indicates a normal police force location or \code{BTP} which indicates a British Transport Police location}
\item{month}{the month of the crime - all 03 (March)}
\item{year}{the year of the crime - all 2021}
}
}
\source{
Created by Lex Comber
}
\usage{
crimes
}
\description{
A point dataset created from data collected from the police API for the Leeds area (see \url{https://data.police.uk} and \url{https://data.police.uk/docs/method/crimes-at-location/})
}
\examples{
data(crimes)
}
\keyword{datasets}
