% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ew_out}
\alias{ew_out}
\title{Outline of England and Wales}
\format{
A \code{sfc} format MultiPolygon in OSGB CRS (EPSG: 27700)
}
\source{
Created by Lex Comber
}
\usage{
ew_out
}
\description{
An outline of the England and Wales area from GWmodel
}
\examples{
data(ew_out)
}
\keyword{datasets}
