% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ewhp}
\alias{ewhp}
\title{House price data for England and Wales}
\format{
A \code{data.frame} with 519 observations and 12 fields
\describe{
\item{Easting}{X coordinate in OSGB projection (EPSG:27700)}
\item{Northing}{Y coordinate in OSGB projection (EPSG:27700)}
\item{PurPrice}{the purchase price of the property}
\item{BldIntWr}{binary indicator, 1 if the property was built during the world war, 0 otherwise}
\item{BldPostW}{binary indicator, 1 if the property was built after the world war, 0 otherwise}
\item{Bld60s}{binary indicator, 1 if the property was built between 1960 and 1969, 0 otherwise}
\item{Bld70s}{binary indicator, 1 if the property was built between 1970 and 1979, 0 otherwise}
\item{Bld80s}{binary indicator, 1 if the property was built between 1980 and 1989, 0 otherwise}
\item{TypDetch}{binary indicator, 1 if the property is detached (i.e. it is a stand-alone house), 0 otherwise}
\item{TypSemiD}{binary indicator, 1 if the property is semi detached, 0 otherwise}
\item{TypFlat}{binary indicator, 1, if the property is a flat (apartment), 0 otherwise}
\item{FlrArea}{floor area of the property in square metres}
}
}
\source{
Created by Chris Brunsdon
}
\usage{
ewhp
}
\description{
A house price data set for England and Wales from 2001 with location and 9 hedonic (explanatory) variables.
}
\examples{
data(ewhp)
}
\keyword{datasets}
