\name{Intro2R4SpatAnal-package}
\alias{Intro2R4SpatAnal-package}
\docType{package}
\title{A Data package for An Introduction to R for Spatial Analysis and Mapping}
\description{
The `Intro2R4SpatAnal` provides data for use in the 3rd edition of the book, 'An Introduction to R for Spatial Analysis and Mapping' by Chirs Brunsdon and Lex Comber.
}
\details{
The `Intro2R4SpatAnal` contains 16 datasets in different formats - spatial and non-spatial - that are used to support examples and apllications in the different chapters of the 3rd edition of the book. All spatial data are in `sf` format.
}
\author{
Lex Comber \email{a.comber@leeds.ac.uk}
}
\keyword{package}
