% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leeds_lsoa}
\alias{leeds_lsoa}
\title{Spatial data of socio-economic attributes from the 2021 UK Census for Leeds, UK}
\format{
A \code{sf} format MULTIPOLYGON dataset with 488 observations (LSOAs) and 24 attributes
\describe{
\item{code}{the LSOA area code}
\item{LSOA21NM}{the LSOA neamed code}
\item{totpop}{the total population count}
\item{u20}{the \% of the population under 20 years old}
\item{o65}{the \% of the population over 65 years old}
\item{badhealth}{the \% of the population reporting the health to be 'bad' or 'very bad'}
\item{nocar}{the \% of households without a car or van}
\item{overcrowd}{the \% of households with insifficient rooms for the occupants}
\item{degree}{the \% of the population with a bachelors degree}
\item{manual}{the \% of the population working in manual occupations (skilled trades occupations;, process, plant and machine operatives and elementary occupations)}
\item{carework}{the \% of the population working in caring, leisure and other service occupations}
\item{owned}{the \% of households owner occupied }
\item{unemp}{the \% of the economicaly active population that is unemployed}
\item{students}{the \% of the population that are students}
\item{indian}{the \% of the population that indicated their ethnicity to be Indian in the census}
\item{pakistani}{the \% of the population that indicated their ethnicity to be Pakistani in the census}
\item{african}{the \% of the population that indicated their ethnicity to be African in the census}
\item{caribbean}{the \% of the population that indicated their ethnicity to be Caribbean in the census}
\item{meg}{the \% of the population that indicated their ethnicity to be multiple ethnic groups in the census}
\item{white}{the \% of the population that indicated their ethnicity to be White in the census}
\item{christian}{the \% of the population that indicated their Religion to be Christian in the census}
\item{hindu}{the \% of the population that indicated their Religion to be Hindu in the census}
\item{jewish}{the \% of the population that indicated their Religion to be Jewish in the census}
\item{muslim}{the \% of the population that indicated their Religion to be Muslim in the census}
\item{sikh}{the \% of the population that indicated their Religion to be Sikh in the census}
\item{incomer1yr}{the \% of the population that moved to the LSOA within the lasyt year (from inside or outsie the UK)}
}
}
\source{
Created by Lex Comber
}
\usage{
leeds_lsoa
}
\description{
A dataset over Lower Super Output Areas (LSOAS), census reporting areas with ~1500 people, (~500 households), for the Leeds city area generated from a selection of the census Topic Summaries (\url{https://www.nomisweb.co.uk/sources/census_2021_ts})
}
\examples{
data(leeds_lsoa)
}
\keyword{datasets}
