% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lsoa_attribs}
\alias{lsoa_attribs}
\title{A data table of 2021 census attributes for LSOAs in Leeds, UK}
\format{
A \code{data.frame} format dataset with 488 observations and 10 fields
\describe{
\item{code}{the LSOA area code}
\item{disabled}{the \% of the population that are disabled}
\item{bornuk}{the \% of the population born in the UK}
\item{metro}{the \% of the population who commute to work by tram or metro }
\item{train}{the \% of the population who commute to work by train }
\item{bus}{the \% of the population who commute to work by bus }
\item{car}{the \% of the population who commute to work by car }
\item{bike}{the \% of the population who commute to work by bicycle }
\item{mbike}{the \% of the population who commute to work by motorbike or scooter }
\item{walk}{the \% of the population who commute to work on foot }
}
devtools::check()
}
\source{
Created by Lex Comber
}
\usage{
lsoa_attribs
}
\description{
A dataset over Lower Super Output Areas (LSOAS), census reporting areas with ~1500 people, (~500 households), for the Leeds city area generated from a selection of the census Topic Summaries (\url{https://www.nomisweb.co.uk/sources/census_2021_ts})
}
\examples{
data(lsoa_attribs)
}
\keyword{datasets}
