% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuts2_unemp_de}
\alias{nuts2_unemp_de}
\title{Unemployment levels in Germany over NUTS2 regions}
\format{
A \code{sf} format Polygon dataset with 38 areas and 3 fields in WGS 84 / Pseudo-Mercator CRS (EPSG: 3857)
\describe{
\item{NUTS_ID}{the NUTS 2 ID}
\item{NAME_LATN}{the NUTS 2 area name}
\item{Unemployment}{the percentage of the working population unemployed in 2018}
}
}
\source{
Created by Chris Brunsdon
}
\usage{
nuts2_unemp_de
}
\description{
A dataset of Unemployment levels in Germany over NUTS2 regions obtained from Eurostat
}
\examples{
data(nuts2_unemp_de)
}
\keyword{datasets}
