% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places}
\alias{places}
\title{A data table of places in and around Leeds, UK}
\format{
A \code{data.frame} format dataset with 12 observations and 3 fields
\describe{
\item{town}{the place name}
\item{lon}{the longitude of the place}
\item{lat}{the latitude of the place}
}
}
\source{
Created by Lex Comber
}
\usage{
places
}
\description{
The names and locations of towns and cities around Leeds, UK
}
\examples{
data(places)
}
\keyword{datasets}
