% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pm10.grid}
\alias{pm10.grid}
\title{A dataset of average annual PM10 for the Leeds area over a 1km grid}
\format{
A \code{data.frame} format dataset with 551 observations and 4 fields
\describe{
\item{X}{observation Easting in metres in OSGB projection (EPSG: 27700)}
\item{Y}{observation Northing in metres in OSGB projection (EPSG: 27700)}
\item{gridcode}{the Ordnance Survey 1km grid code}
\item{pm102021g}{the mean PM10 level in µg m^3}
}
}
\source{
Created by Lex Comber
}
\usage{
pm10.grid
}
\description{
A dataset of the mean PM10 levels in 2021 from \url{https://uk-air.defra.gov.uk/data/pcm-data}
}
\examples{
data(pm10.grid)
}
\keyword{datasets}
