% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pm10}
\alias{pm10}
\title{A 1km gridded dataset of average annual PM10 for the Leeds area}
\format{
A \code{sf} format POINT dataset with 551 observations and 2 fields in OSGB projection (EPSG: 27700)
\describe{
\item{gridcode}{the Ordnance Survey 1km grid code}
\item{pm102021g}{the mean PM10 level in µg m^3}
}
}
\source{
Created by Lex Comber
}
\usage{
pm10
}
\description{
A point dataset of the mean PM10 levels in 2021 from \url{https://uk-air.defra.gov.uk/data/pcm-data}
}
\examples{
data(pm10)
}
\keyword{datasets}
