% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roads}
\alias{roads}
\title{Road data for the Leeds area from OpenStreetMap}
\format{
A \code{sf} format LINESTRING dataset with 8691 observations and 4 fields  in WGS84 projection (EPSG: 4326)
\describe{
\item{name}{the name of the road where indicated in OSM}
\item{highway}{the class of road (primary, secondary, teriary and trunk)}
\item{ref}{the road number where indicated in OSM}
\item{length}{the road segment length in metres}
}
}
\source{
Created by Lex Comber
}
\usage{
roads
}
\description{
A dataset of the main roads in and around the Leeds areas in the UK, collected from OpenStreetMap in April 2024
}
\examples{
data(roads)
}
\keyword{datasets}
