% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{docs}
\alias{docs}
\title{Doctor surgery locations in Leeds from OpenStreetMap}
\format{
A \code{sf} format POINT dataset with 987 observations and 1 fields in OSGB projection (EPSG: 27700)
\describe{
\item{name}{the name of the doctor surgery where present in OSM}
}
}
\source{
Created by Lex Comber
}
\usage{
docs
}
\description{
A point dataset of doctor surgeries in Leeds from OpenStreetMap, downloaded on 25 April 2024
}
\examples{
data(docs)
}
\keyword{datasets}
