% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lads.polys}
\alias{lads.polys}
\title{Data of the outlines of Local Authority Districts in the UK}
\format{
A named list object with 363 elements, named by the LAD
\describe{
\item{X}{the X coordinates of the boundary in OSGB projection (EPSG: 27700)}
\item{y}{the Y coordinates of the boundary in OSGB projection (EPSG: 27700)}
}
}
\source{
Created by Lex Comber
}
\usage{
lads.polys
}
\description{
A dataset containing the boundaries of the LADs in the UK (\url{https://geoportal.statistics.gov.uk/documents/d1fab2d9fb0a4576a7e08f89ac7e0b72/about})
}
\examples{
data(lads.polys)
}
\keyword{datasets}
